codigo final

+++ R2D2/inc/PR_timer.h
@@ -18,14 +18,14 @@
 typedef void (*Timer_Handler)(void);
 
 void initTimer(void);
-uint8_t startTimer(uint32_t time, Timer_Closure handler , uint8_t base );
-void killTimer(uint8_t descriptor);
+uint16_t startTimer(uint32_t time, Timer_Closure handler , uint8_t base );
+void killTimer(uint16_t descriptor);
 void fastTimer(uint32_t time, Timer_Closure handler);
 void wait(uint32_t time);
-uint8_t startCronometer();
-uint32_t stopCronometer(uint8_t descriptor);
-uint32_t getCronometer(uint8_t descriptor);
-uint32_t getCronometerInBase(uint8_t descriptor, uint8_t base);
-uint32_t stopCronometerInBase(uint8_t descriptor, uint8_t base);
+uint16_t startCronometer();
+uint32_t stopCronometer(uint16_t descriptor);
+uint32_t getCronometer(uint16_t descriptor);
+uint32_t getCronometerInBase(uint16_t descriptor, uint8_t base);
+uint32_t stopCronometerInBase(uint16_t descriptor, uint8_t base);
 
 #endif /* HEADERS_PR_TIMER_H_ */
